backend/requirements.txt
fastapi
uvicorn[standard]
web3
python-dotenv
firebase-admin
httpx
pydantic
sqlalchemy
psycopg2-binary
email-validator
